requirements.txt
conda==3.19.0
conda-env==2.4.5
csvkit==0.9.1
cycler==0.9.0
dbf==0.96.3
dstools==0.3
enum34==1.1.1
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
lxml==3.5.0
MarkupSafe==0.23
matplotlib==1.5.0
numpy==1.10.2
openpyxl==1.8.5
pandas==0.17.1
psycopg2==2.6.1
pyaml==15.8.2
pycosat==0.6.1
pycrypto==2.6.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.9.0
scikit-learn==0.17
scipy==0.16.1
shyaml==0.3.4
six==1.10.0
SQLAlchemy==1.0.10
Werkzeug==0.11.3
wheel==0.26.0
xlrd==0.9.4
xlsx2csv==0.7.2
sklearn-evaluation==0.1